phyzzie first commit
this is the beginning of a simple framework for designing and generating
simulations of physical systems defined by static json world
descriptions. phyzzie uses the p2.js physics engine as a back end.

the intended purpose of this framework is to facilitate the rapid
development of experiments which evolve artificial neural networks.