don't reduce sim time in simulation mode
if the outer layer of the library passed a time delta greater than the
max time step, then an inner layer would silently simulate less time.

for the sake of avoiding issues should the outer layer change, i've
refactored the code so that time shortening only occurs explicitly in
that outer layer when rendering to the screen in real time.